backend/requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# PyTorch — installed separately in Dockerfile (CUDA) or via pip (CPU)
# Do NOT list torch/torchvision here to avoid double-install in Docker.
# For local dev: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121

# MongoDB (async driver) — used when USE_MONGODB=true
motor>=3.3.0
pymongo[srv]>=4.6.0
certifi>=2023.0.0

# SQLite (async driver) — used when USE_MONGODB=false (test mode)
aiosqlite>=0.19.0

# AI & ML
fer>=0.1.4
google-generativeai>=0.3.2
pillow>=10.2.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
mediapipe>=0.10.9

# Validation & Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0        # required by Pydantic EmailStr
python-dotenv>=1.0.0

# Speech-to-Text (ASR)
openai-whisper>=20231117

# Utilities
aiofiles>=23.2.1

# Authentication
python-jose[cryptography]>=3.3.0  # JWT tokens
passlib[bcrypt]>=1.7.4            # Password hashing via CryptContext
bcrypt>=4.1.0
